Repurpose spin-sqlite crate
