+++ crates/factor-sqlite/src/runtime_config.rs
@@ -1,6 +1,3 @@
-#[cfg(feature = "spin-cli")]
-pub mod spin;
-
 use std::{collections::HashMap, sync::Arc};
 
 use crate::ConnectionCreator;
