+++ crates/runtime-config/Cargo.toml
@@ -25,6 +25,7 @@ spin-factor-sqlite = { path = "../factor-sqlite" }
 spin-factor-variables = { path = "../factor-variables" }
 spin-factor-wasi = { path = "../factor-wasi" }
 spin-factors = { path = "../factors" }
+spin-sqlite = { path = "../sqlite" }
 toml = "0.8"
 
 [lints]
