+++ crates/sqlite-inproc/Cargo.toml
@@ -6,11 +6,11 @@ edition = { workspace = true }
 
 [dependencies]
 anyhow = "1.0"
-async-trait = "0.1.68"
+async-trait = "0.1"
 once_cell = "1"
 rand = "0.8"
 rusqlite = { version = "0.29.0", features = ["bundled"] }
-spin-sqlite = { path = "../sqlite" }
+spin-factor-sqlite = { path = "../factor-sqlite" }
 spin-world = { path = "../world" }
 tokio = "1"
 tracing = { workspace = true }
